Can now handle invalid body/cookie values
